+++ src/Sentry/SentrySdk.cs
@@ -521,6 +521,21 @@ public static partial class SentrySdk
         => CurrentHub.CaptureSession(sessionUpdate);
 
     /// <summary>
+    /// Captures a check-in.
+    /// </summary>
+    /// <remarks>
+    /// Capturing a check-in returns an ID. The ID can be used to update the status. I.e. to update a check-in you
+    /// captured from `CheckInStatus.InProgress` to `CheckInStatus.Ok`.
+    /// </remarks>
+    /// <param name="monitorSlug">The monitor slug of the check-in.</param>
+    /// <param name="status">The status of the check-in.</param>
+    /// <param name="sentryId">The optional <see cref="SentryId"/>.</param>
+    /// <returns>The Id of the check-in.</returns>
+    [DebuggerStepThrough]
+    public static SentryId CaptureCheckIn(string monitorSlug, CheckInStatus status, SentryId? sentryId)
+        => CurrentHub.CaptureCheckIn(monitorSlug, status, sentryId);
+
+    /// <summary>
     /// Starts a transaction.
     /// </summary>
     [DebuggerStepThrough]